승필 - 페이징처리와 채팅 리팩토링 일부분, 네고

+++ final_TresureStore/src/main/java/com/kh/tresure/member/model/vo/Account.java
@@ -13,7 +13,7 @@ public class Account {
 
 	private int accountId; //계좌 고유 식별번호
 	private int userNo; //유저번호
-	private long account; //계좌번호
+	private int account; //계좌번호
 	
 	
 }
